Moved _exceptions into exceptions
