+++ stingray/exceptions.py
